Collapse opfs-btree into one module
