+++ crates/opfs-btree/src/lib.rs
@@ -1,5 +1,4 @@
 mod db;
-mod durable;
 mod error;
 mod file;
 mod page;
@@ -7,8 +6,7 @@ mod superblock;
 #[cfg(target_arch = "wasm32")]
 pub mod wasm_bench;
 
-pub use db::BTreeOptions;
-pub use durable::{OpfsBTree, OpfsBTreeFiles};
+pub use db::{BTreeOptions, OpfsBTree, OpfsBTreeFiles};
 pub use error::BTreeError;
 #[cfg(target_arch = "wasm32")]
 pub use file::OpfsFile;
